move `TestServer` into common and expose only `api_server`
