+++ tests/integration_test.rs
@@ -1,7 +1,2 @@
 mod api;
-
-#[test]
-fn it_add_something() {
-    let x = 1 + 2;
-    assert_eq!(3, x)
-}
+mod common;
